endpoints/requirements.txt
fastapi==0.95.2
sse-starlette
uvicorn[standard]==0.21.1
pydantic
python-dotenv==0.15.0
boto3==1.26.148
requests==2.28.0

# Extra
langchain==0.1.5
chromadb==0.4.22

openai==1.3.7
tiktoken
google-generativeai==0.3.1
